Cell object
Created Cell object to store each cell in the grid, along with its coordinates.
This, along with the get_all method to get all the rows, column and squares, allowed a great simplification of various methods, such as "fill_missing_single_rows" and its variants, which are now condensed into "fill_missing_single".