Added pause, previous and next
Added buttons and functionality to pause/resume the solve, show the previous step, and move along the stages of the grid seamlessly.
There seems to be a bug where a number that was zero is restored to the latest found number when going back a few steps. For example, currently a square will contain 12, then pressing previous will change it to 0, but pressing again will bring the 12 back. Going back further restores the 0.