Removed single step pause
Removed all pauses after a single step forward, so now multiple single steps forward can be made right after each other.
It's not as responsive as I'd like, but unless the button is pressed really fast there are no issues. This is almost definitely not caused by the time_between_steps since the issue is on the main thread and freezes the UI for almost a second sometimes. It seems to be more noticeable deeper in the solve.