Fixed candidate groups
Fixed an issue where candidates found inside a candidate group would be removed from the members of the group, rendering the strategy useless. It is unclear why the strategy still gave correct-looking results before. Now, the strategy is never used successfully in testing, and the puzzle at index 5 can no longer be completed.
To investigate this, I will show all candidates for each square in the explanation in case the puzzle wasn't finished.