Implemented subset combinations
The subsets chosen to find disjoint subsets are now composed of any combination of squares from the empty squares in the sequence.
Before, the subsets were only found if the shared candidates were among the first candidates in the square. The previous implementation also didn't check all possible combinations of empty squares, which is now done using a bitmask.